requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==25.1.0
bech32==1.2.0
betterproto==2.0.0b5
bip32utils==0.3.post4
blinker==1.9.0
boltons==21.0.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.16.1
email-validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozenlist==1.5.0
furl==2.1.3
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
miscreant==0.3.0
mnemonic==0.19
multidict==6.1.0
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==2.2.0
ollama==0.4.4
orderedmultidict==1.0.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==3.20.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
secp256k1==0.14.0
secret-ai-sdk==0.1.2
secret-sdk==1.8.1
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.45.3
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0
wheel==0.44.0
wrapt==1.17.0
yarl==1.18.3
